moved backtracking algorithms to backtracking dir, added subskills , fixed readme
