+++ src/main/java/leetcode/Backtracking/Combinations/Combinations.java
@@ -1,4 +1,27 @@
-package leetcode.Combinations;
+package leetcode.Backtracking.Combinations;
+
+/***
+ * medium 77
+ * 
+ * Given two integers n and k, return all possible combinations of k numbers chosen from the range [1, n].
+ * You may return the answer in any order.
+ * 
+ * 
+ * Example 1:
+ * Input: n = 4, k = 2
+ * Output: [[1,2],[1,3],[1,4],[2,3],[2,4],[3,4]]
+ * Explanation: There are 4 choose 2 = 6 total combinations.
+ * Note that combinations are unordered, i.e., [1,2] and [2,1] are considered to be the same combination.
+ * 
+ * 
+ * Example 2:
+ * Input: n = 1, k = 1
+ * Output: [[1]]
+ * Explanation: There is 1 choose 1 = 1 total combination.
+ * 
+ * T.C.: O(n^2);
+ * S.C.: O(2n).
+ */
 
 import java.util.ArrayList;
 import java.util.List;
