+++ src/main/java/leetcode/Backtracking/LetterCombinationsOfPhoneNumber/LetterCombinationsOfPhoneNumber.java
@@ -1,4 +1,23 @@
-package leetcode.LetterCombinationsOfPhoneNumber;
+package leetcode.Backtracking.LetterCombinationsOfPhoneNumber;
+
+/***
+ * medium 17
+ * 
+ * Example 1:
+ * Input: digits = "23"
+ * Output: ["ad","ae","af","bd","be","bf","cd","ce","cf"]
+ * 
+ * Example 2:
+ * Input: digits = ""
+ * Output: []
+ * 
+ * Example 3:
+ * Input: digits = "2"
+ * Output: ["a","b","c"]
+ * 
+ * T.C.: O(n)
+ * S.C.: O(n)
+ */
 
 import java.util.ArrayList;
 import java.util.Hashtable;
